Added a fixture to disable xcvrd
Details:
As part of sfp utils reset test, I added a fixture to stop xcvrd in pmon service during the test,
to avoid race condition when accessing the transceiver. After the test the fixture enables xcvrd again.
This fixture replaces the LogAnalyzer skips to error messages that currently appear becuase the cable EEPROM is not accessible
In the suggested flow, these error messages won't appear.