[console] Optimize test_console_baud_rate (#8370)
What is the motivation for this PR?
If we are requested to enter username and pwd for DUT, it means that we connected to DUT from console successfully. So, it is no need to login DUT from console.

How did you do it?
Remove login step.

How did you verify/test it?
Run test.