[Snappi] Validate the IPinIP packet can be transferred without any re-order (#8202)
Summary: We observed some IPinIP packets of various sizes are reordered when passing through a switch even though they belong to the same stream. As a result, this test case aims to fill this test gap by passing IPinIp packets between two ports on a traffic generator through a switch to see if the packets get re-ordered by the switch.

Fixes (issue) #7525

What is the motivation for this PR?
There was a customer issue in which a switch was dropping IPinIP packets of varying sizes, and no test cases to cover this scenario in which the packets were re-ordered causing customer impact. Hence, this test case aims to check for packet re-order scenarios in switches.

How did you do it?
Generate a sequence of packet and flow configurations and run the traffic from a traffic generator through the DUT (device under test) to see if the packets are getting re-ordered.

Here is the general topo:
traffic gen(port 1) --- dut ---- traffic gen(port 2)

How did you verify/test it?
Ran the test on the affected switch in the lab and see that it fails for a specific software version, and passes after the software fix is introduced.