Improve report_uploader to upload files by best effort (#8504)
What is the motivation for this PR?
The report_uploader.py script supports uploading multiple report files in a batch to Kusto. However, if any of the report file is malformed, an exception would be raised and the entire uploading will fail.

How did you do it?
This change improved the report_uploader.py script to support uploading reports by best effort. This means that if any report file is malformed, that file will be skipped. The good report files still can be processed and uploaded.

How did you verify/test it?
Tested uploading multiple report files. One of the files is a malformed json file.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>