Fit dualtor scenario when deleting config VLAN|mac in module vlan/test_host_vlan.py. (#9025)
Description of PR
In PR #8728, we delete the vlan mac running config in the teardown period. But in dualtor scenario, it already has mac address in original config. So in this PR, when the testbed is dualtor, we skip deleting the vlan mac running config in the teardown period.

What is the motivation for this PR?
In PR #8728, we delete the vlan mac running config in the teardown period. But in dualtor scenario, it already has mac address in original config. So in this PR, when the testbed is dualtor, we skip deleting the vlan mac running config in the teardown period.

How did you do it?
Skip deleting the vlan mac running config in the teardown period when the testbed is dualtor.

How did you verify/test it?
```
08:25:47 conftest.core_dump_and_config_check      L2131 INFO   | Core dump and config check passed for vlan/test_host_vlan.py
```

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>