Recover telemetry settings when error happens in module telemetry/test_telemetry.py (#9173)
Description of PR
In module telemetry/test_telemetry.py, errors may happen in the setup period of fixture setup_streaming_telemetry, and the the recover function can be not be executed in the teardown period. This will change the running config, and cause unnecessary reload. In this PR, when the errors happen in the setup period, we will recover the telemetry config immediately.

What is the motivation for this PR?
In module telemetry/test_telemetry.py, errors may happen in the setup period of fixture setup_streaming_telemetry, and the the recover function can be not be executed in the teardown period. This will change the running config, and cause unnecessary reload. In this PR, when the errors happen in the setup period, we will recover the telemetry config immediately.

How did you do it?
When the errors happen in the setup period, we will recover the telemetry config immediately.

How did you verify/test it?
07:35:09 test_telemetry.setup_streaming_telemetry L0142 INFO   | Error happens in the setup period of setup_streaming_telemetry, recover the telemetry.
07:36:35 conftest.core_dump_and_config_check      L1890 INFO   | Core dump and config check passed for telemetry/test_telemetry.py

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>