[tests/crm] Fix python3 syntax error in `test_acl_entry` (#9208)
What is the motivation for this PR?
Fix a python3 syntax error in testcase crm/test_crm.py::test_acl_entry. In python2 dict.keys() returns a list object, but in python3 it returns a dict_keys object which is not sortable.

How did you do it?
Use sorted() function instead of dict.keys().sort(). sorted() function can put the keys of a dict in a list in sorted order, and it has same action in Python2 and Python3.

How did you verify/test it?
Verified by running testcase on Marvell testbed.

Signed-off-by: Zhijian Li <zhijianli@microsoft.com>