Revert the change to template acl.json.  (#9346)
Description of PR
PR #9199 modified the template acl.json under folder tests/crm/templates and added a key input_interface in this template. Only backend topo supports this key, so it will generate an error log in non-backend topo. In this PR, I revert the change to this template. And for recovery, we have another template in PR #9312 .

What is the motivation for this PR?
PR #9199 modified the template acl.json under folder tests/crm/templates and added a key input_interface in this template. Only backend topo supports this key, so it will generate an error log in non-backend topo. In this PR, I revert the change to this template. And for recovery, we have another template in PR #9312

How did you do it?
Revert the change to template acl json.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>