[Testbed] Fix the MAC address of br1 on server during add/remove topo (#9877)
Summary:
Fix the MAC address of br1 on server during add/remove topo.

As we configure the MGMT IP address of server on a virtual bridge br1, it's MAC address is not fixed if it's not manually set. This makes the MAC address of br1 changeable during add/remove topo. When the MAC address change, the other servers may not update their ARP table in time and temporarily lose the connection. In this PR, I added one step in add/remove topo to fix the MAC address of br1 to avoid potential packet loss.

What is the motivation for this PR?
Fix the MAC address of br1 on server during add/remove topo.

How did you do it?
Add one step to explicitly set the MAC address of br1.

How did you verify/test it?
* For physical testbeds, verified on physical servers.
* For virtual testbeds, verified by PR test.